:refactor:(useTrekFilter) Define hook
